pyproject.toml
[project]
name = "briefbot"
version = "2.0.0"
description = "BriefBot — AI assistant for processing client voice/text messages into PDF briefs"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.4",
    "openai>=1.30",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "supabase>=2.4",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "structlog>=24.1",
    "httpx>=0.27",
    "aiofiles>=23.2",
    "tenacity>=8.2",
    "packaging>=23.0",
    "google-genai>=1.0.0",
    "groq>=0.9.0",
    "jinja2>=3.1",
    "weasyprint>=62",
    "uvloop>=0.19; sys_platform != 'win32'",
    "psycopg2-binary>=2.9",
    "mistune>=3.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "websockets>=12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
